混合型支持映射型代码调整

+++ metric-calculate-config/src/main/java/com/yanggu/metric_calculate/config/service/impl/MixUdafParamItemServiceImpl.java
@@ -1,19 +1,18 @@
 package com.yanggu.metric_calculate.config.service.impl;
 
+import com.mybatisflex.core.query.QueryWrapper;
 import com.mybatisflex.spring.service.impl.ServiceImpl;
 import com.yanggu.metric_calculate.config.mapper.MixUdafParamItemMapper;
-import com.yanggu.metric_calculate.config.pojo.entity.BaseUdafParam;
-import com.yanggu.metric_calculate.config.pojo.entity.MapUdafParam;
-import com.yanggu.metric_calculate.config.pojo.entity.MixUdafParamItem;
-import com.yanggu.metric_calculate.config.pojo.entity.ModelColumn;
-import com.yanggu.metric_calculate.config.service.BaseUdafParamService;
-import com.yanggu.metric_calculate.config.service.MapUdafParamService;
-import com.yanggu.metric_calculate.config.service.MixUdafParamItemService;
+import com.yanggu.metric_calculate.config.pojo.entity.*;
+import com.yanggu.metric_calculate.config.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static com.yanggu.metric_calculate.config.pojo.entity.table.MixUdafParamItemBaseUdafParamRelationTableDef.MIX_UDAF_PARAM_ITEM_BASE_UDAF_PARAM_RELATION;
+import static com.yanggu.metric_calculate.config.pojo.entity.table.MixUdafParamItemMapUdafParamRelationTableDef.MIX_UDAF_PARAM_ITEM_MAP_UDAF_PARAM_RELATION;
 
 /**
  * 混合聚合参数，混合聚合类型定义。value只能是数值型、集合型、对象型中间表 服务层实现。
@@ -27,6 +26,12 @@ public class MixUdafParamItemServiceImpl extends ServiceImpl<MixUdafParamItemMap
     @Autowired
     private MapUdafParamService mapUdafParamService;
 
+    @Autowired
+    private MixUdafParamItemBaseUdafParamRelationService mixUdafParamItemBaseUdafParamRelationService;
+
+    @Autowired
+    private MixUdafParamItemMapUdafParamRelationService mixUdafParamItemMapUdafParamRelationService;
+
     @Override
     @Transactional(rollbackFor = RuntimeException.class)
     public void saveData(MixUdafParamItem mixUdafParamItem, List<ModelColumn> modelColumnList) throws Exception {
@@ -36,22 +41,44 @@ public class MixUdafParamItemServiceImpl extends ServiceImpl<MixUdafParamItemMap
         if (baseUdafParam != null) {
             baseUdafParamService.saveData(baseUdafParam, modelColumnList);
             //保存中间表
+            MixUdafParamItemBaseUdafParamRelation relation = new MixUdafParamItemBaseUdafParamRelation();
+            relation.setMixUdafParamItemId(mixUdafParamItemId);
+            relation.setBaseUdafParamId(baseUdafParam.getId());
+            mixUdafParamItemBaseUdafParamRelationService.save(relation);
         }
         MapUdafParam mapUdafParam = mixUdafParamItem.getMapUdafParam();
         if (mapUdafParam != null) {
             mapUdafParamService.saveData(mapUdafParam, modelColumnList);
             //保存中间表
+            MixUdafParamItemMapUdafParamRelation relation = new MixUdafParamItemMapUdafParamRelation();
+            relation.setMixUdafParamItemId(mixUdafParamItemId);
+            relation.setMapUdafParamId(mapUdafParam.getId());
+            mixUdafParamItemMapUdafParamRelationService.save(relation);
         }
     }
 
     @Override
     @Transactional(rollbackFor = RuntimeException.class)
     public void deleteData(MixUdafParamItem mixUdafParamItem) {
-        Integer id = mixUdafParamItem.getId();
-        super.removeById(id);
+        Integer mixUdafParamItemId = mixUdafParamItem.getId();
+        super.removeById(mixUdafParamItemId);
         //删除中间表和数据
         BaseUdafParam baseUdafParam = mixUdafParamItem.getBaseUdafParam();
-        baseUdafParamService.deleteData(baseUdafParam);
+        if (baseUdafParam != null) {
+            baseUdafParamService.deleteData(baseUdafParam);
+            QueryWrapper queryWrapper = QueryWrapper.create()
+                    .where(MIX_UDAF_PARAM_ITEM_BASE_UDAF_PARAM_RELATION.MIX_UDAF_PARAM_ITEM_ID.eq(mixUdafParamItemId))
+                    .and(MIX_UDAF_PARAM_ITEM_BASE_UDAF_PARAM_RELATION.BASE_UDAF_PARAM_ID.eq(baseUdafParam.getId()));
+            mixUdafParamItemBaseUdafParamRelationService.remove(queryWrapper);
+        }
+        MapUdafParam mapUdafParam = mixUdafParamItem.getMapUdafParam();
+        if (mapUdafParam != null) {
+            mapUdafParamService.deleteData(mapUdafParam);
+            QueryWrapper queryWrapper = QueryWrapper.create()
+                    .where(MIX_UDAF_PARAM_ITEM_MAP_UDAF_PARAM_RELATION.MIX_UDAF_PARAM_ITEM_ID.eq(mixUdafParamItemId))
+                    .and(MIX_UDAF_PARAM_ITEM_MAP_UDAF_PARAM_RELATION.MAP_UDAF_PARAM_ID.eq(mapUdafParam.getId()));
+            mixUdafParamItemMapUdafParamRelationService.remove(queryWrapper);
+        }
     }
 
 }
